fixed resume route to keep route.ts HTTP methods to single file per method to avoid cache issues
